Continous refactoring of using Upercase for Classes
