+++ topojson/__init__.py
@@ -1,10 +1,24 @@
 __version__ = "1.0rc2"
 __doc__ = "Encode geographic data as topology in Python! 🌍"
 
-# from .extract import Extract
-# from .join import Join
-# from .cut import Cut
-# from .dedup import Dedup
-# from .hashmap import Hashmap
-from .topology import Topology
+
+from .core.extract import Extract
+from .core.join import Join
+from .core.cut import Cut
+from .core.dedup import Dedup
+from .core.hashmap import Hashmap
+from .core.topology import Topology
+
+
+def topology(data):
+    import warnings
+
+    warnings.warn(
+        (
+            "\nThe function topojson.topology() is deprecated.\n\n"
+            "You have to use topojson.Topology() since v1.0rc3."
+        ),
+        DeprecationWarning,
+        stacklevel=2,
+    )
