+++ topojson/core/hashmap.py
@@ -2,6 +2,7 @@ import numpy as np
 from itertools import compress
 from simplification import cutil
 import copy
+import pprint
 from .dedup import Dedup
 
 
@@ -18,7 +19,109 @@ class Hashmap(Dedup):
         self.simp = False
 
         # execute main function of Hashmap
-        self.hashmap = self.hashmapper(self.dedup)
+        self.output = self.hashmapper(self.output)
+
+    def __repr__(self):
+        return "Hashmap(\n{}\n)".format(pprint.pformat(self.output))
+
+    def to_dict(self):
+        return self.output
+
+    def hashmapper(self, data, simplify_factor=None):
+        """
+        Hashmap function resolves bookkeeping results to object arcs.
+
+        The hashmap function is the fifth step in the topology computation.
+        The following sequence is adopted:
+        1. extract
+        2. join
+        3. cut 
+        4. dedup   
+        5. hashmap   
+ 
+        Developping Notes:
+        * PostGIS Tips for Power Users: http://2010.foss4g.org/presentations/3369.pdf
+        """
+
+        # make data available within class
+        self.data = data
+
+        # resolve bookkeeping to arcs in objects, including backward check of arcs
+        list(self.resolve_objects("arcs", self.data["objects"]))
+
+        # parse the linestrings into list of coordinates
+        # only if linestrings are quantized, apply delta encoding.
+        if "transform" in data.keys():
+            if simplify_factor is not None:
+                if simplify_factor >= 1:
+                    for idx, ls in enumerate(data["linestrings"]):
+                        self.data["linestrings"][idx] = cutil.simplify_coords(
+                            np.array(ls), simplify_factor
+                        )
+                    self.simp = True
+
+            for idx, ls in enumerate(data["linestrings"]):
+                if self.simp:
+                    ls = ls.astype(int)
+                else:
+                    ls = np.array(ls).astype(int)
+                ls_p1 = copy.copy(ls[0])
+                ls -= np.roll(ls, 1, axis=0)
+                ls[0] = ls_p1
+                self.data["linestrings"][idx] = ls.tolist()
+
+        else:
+            if simplify_factor is not None:
+                if simplify_factor >= 1:
+                    for idx, ls in enumerate(data["linestrings"]):
+                        self.data["linestrings"][idx] = cutil.simplify_coords(
+                            np.array(ls), simplify_factor
+                        ).tolist()
+            else:
+                for idx, ls in enumerate(data["linestrings"]):
+                    self.data["linestrings"][idx] = np.array(ls).tolist()
+
+        objects = {}
+        objects["geometries"] = []
+        objects["type"] = "GeometryCollection"
+        for idx, feature in enumerate(data["objects"]):
+            feat = data["objects"][feature]
+
+            if "geometries" in feat:
+                feat["type"] = feat["geometries"][0]["type"]
+
+            if feat["type"] == "Polygon":
+                if "geometries" in feat:
+                    f_arc = feat["geometries"][0]["arcs"]
+                else:
+                    f_arc = feat["arcs"]
+
+                feat["arcs"] = f_arc
+
+            if feat["type"] == "MultiPolygon":
+                if "geometries" in feat:
+                    f_arcs = feat["geometries"][0]["arcs"]
+                else:
+                    f_arcs["arcs"]
+                feat["arcs"] = [[arc] for arc in f_arcs]
+
+            feat.pop("geometries", None)
+            objects["geometries"].append(feat)
+
+        data["objects"] = {}
+        data["objects"]["data"] = objects
+
+        # prepare to return object
+        data = self.data
+        data["arcs"] = data["linestrings"]
+        del data["linestrings"]
+        del data["junctions"]
+        del data["bookkeeping_geoms"]
+        del data["bookkeeping_duplicates"]
+        del data["bookkeeping_arcs"]
+        del data["bookkeeping_shared_arcs"]
+
+        return data
 
     def hash_order(self, arc_ids, shared_bool):
         """
@@ -223,106 +326,3 @@ class Hashmap(Dedup):
                 for d in v:
                     for result in self.resolve_objects(key, d):
                         yield result
-
-    def hashmapper(self, data, simplify_factor=None):
-        """
-        Hashmap function resolves bookkeeping results to object arcs.
-
-        The hashmap function is the fifth step in the topology computation.
-        The following sequence is adopted:
-        1. extract
-        2. join
-        3. cut 
-        4. dedup   
-        5. hashmap   
- 
-        Developping Notes:
-        * PostGIS Tips for Power Users: http://2010.foss4g.org/presentations/3369.pdf
-        """
-
-        # make data available within class
-        self.data = data
-
-        # resolve bookkeeping to arcs in objects, including backward check of arcs
-        list(self.resolve_objects("arcs", self.data["objects"]))
-
-        # parse the linestrings into list of coordinates
-        # only if linestrings are quantized, apply delta encoding.
-        if "transform" in data.keys():
-            if simplify_factor is not None:
-                if simplify_factor >= 1:
-                    for idx, ls in enumerate(data["linestrings"]):
-                        self.data["linestrings"][idx] = cutil.simplify_coords(
-                            np.array(ls), simplify_factor
-                        )
-                    self.simp = True
-
-            for idx, ls in enumerate(data["linestrings"]):
-                if self.simp:
-                    ls = ls.astype(int)
-                else:
-                    ls = np.array(ls).astype(int)
-                ls_p1 = copy.copy(ls[0])
-                ls -= np.roll(ls, 1, axis=0)
-                ls[0] = ls_p1
-                self.data["linestrings"][idx] = ls.tolist()
-
-        else:
-            if simplify_factor is not None:
-                print("xxx")
-                if simplify_factor >= 1:
-                    for idx, ls in enumerate(data["linestrings"]):
-                        self.data["linestrings"][idx] = cutil.simplify_coords(
-                            np.array(ls), simplify_factor
-                        ).tolist()
-            else:
-                for idx, ls in enumerate(data["linestrings"]):
-                    self.data["linestrings"][idx] = np.array(ls).tolist()
-
-        objects = {}
-        objects["geometries"] = []
-        objects["type"] = "GeometryCollection"
-        for idx, feature in enumerate(data["objects"]):
-            feat = data["objects"][feature]
-
-            if "geometries" in feat:
-                feat["type"] = feat["geometries"][0]["type"]
-
-            if feat["type"] == "Polygon":
-                if "geometries" in feat:
-                    f_arc = feat["geometries"][0]["arcs"]
-                else:
-                    f_arc = feat["arcs"]
-
-                feat["arcs"] = f_arc
-
-            if feat["type"] == "MultiPolygon":
-                if "geometries" in feat:
-                    f_arcs = feat["geometries"][0]["arcs"]
-                else:
-                    f_arcs["arcs"]
-                feat["arcs"] = [[arc] for arc in f_arcs]
-
-            feat.pop("geometries", None)
-            objects["geometries"].append(feat)
-
-        data["objects"] = {}
-        data["objects"]["data"] = objects
-
-        # prepare to return object
-        data = self.data
-        data["arcs"] = data["linestrings"]
-        del data["linestrings"]
-        del data["junctions"]
-        del data["bookkeeping_geoms"]
-        del data["bookkeeping_duplicates"]
-        del data["bookkeeping_arcs"]
-        del data["bookkeeping_shared_arcs"]
-
-        return data
-
-
-# def hashmap(data, simplify_factor=None):
-#     data = copy.deepcopy(data)
-#     hashmapper = Hashmap()
-#     return hashmapper.main(data, simplify_factor)
